+++ desktop/src-tauri/src/appdata.rs
@@ -3,13 +3,14 @@ use crate::structs::{
 };
 use futures_signals::signal::Mutable;
 use serde::{Deserialize, Serialize};
+use specta::Type;
 use strum::EnumTryAs;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
+#[derive(Debug, Clone, Serialize, Deserialize, Type)]
 pub struct ExperimentState {
     pub experiment: Experiment,
     pub experiment_result: ExperimentResult,
-    pub current_index: usize,
+    pub current_index: u32,
     pub choice_current_preset: CurrentPreset,
 }
 
@@ -38,7 +39,7 @@ impl ExperimentState {
 
         let preset = match &experiment.experiment_type {
             ExperimentType::Choice { choices } => {
-                let choice = &choices[*current_index];
+                let choice = &choices[*current_index as usize];
 
                 match choice_current_preset {
                     CurrentPreset::A => experiment.presets[&choice.a].clone(),
@@ -47,7 +48,7 @@ impl ExperimentState {
             }
 
             ExperimentType::Rating { order } => {
-                let preset_name = &order[*current_index];
+                let preset_name = &order[*current_index as usize];
 
                 experiment.presets[preset_name].clone()
             }
@@ -64,7 +65,7 @@ impl ExperimentState {
     }
 }
 
-#[derive(Debug, Clone, EnumTryAs, Serialize, Deserialize)]
+#[derive(Debug, Clone, EnumTryAs, Serialize, Deserialize, Type)]
 pub enum AppState {
     LiveView(RenderParameters),
     Experiment(ExperimentState),
